+++ vrp-pragmatic/src/construction/enablers/mod.rs
@@ -2,7 +2,7 @@
 
 use std::sync::Arc;
 use vrp_core::models::common::Dimensions;
-use vrp_core::models::problem::Single;
+use vrp_core::models::problem::{Job, Single};
 use vrp_core::models::solution::{Activity, Route};
 
 mod entities;
@@ -35,6 +35,11 @@ pub(crate) fn get_vehicle_id_from_job(job: &Arc<Single>) -> &String {
 pub(crate) fn is_correct_vehicle(route: &Route, target_id: &str, target_shift: usize) -> bool {
     route.actor.vehicle.dimens.get_vehicle_id().expect("cannot get vehicle id") == target_id
         && get_shift_index(&route.actor.vehicle.dimens) == target_shift
+}
+
+pub(crate) fn is_job_belongs_to_route(route: &Route, job: &Job) -> bool {
+    job.as_single()
+        .map_or(false, |job| is_correct_vehicle(route, get_vehicle_id_from_job(job), get_shift_index(&job.dimens)))
 }
 
 pub(crate) fn is_single_belongs_to_route(route: &Route, single: &Arc<Single>) -> bool {