+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Removed
 
 * GitHub sponsors functionality
-* experimetnal `async-evolution` feature
+* experimental `async-evolution` feature
 
 ### Fixed
 
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 * tweak rosomaxa algorithm
 * tweak job index
 * tweak infeasible search heuristic
+* refactor route intervals and multi trip enablers
 
 
 ## [v1.23.0]- 2023-12-22